+++ R/EpiStats.R
@@ -114,6 +114,8 @@ build_epistats <- function(geog.lvl = NULL,
                            init.hiv.prev = NULL,
                            time.unit = 7,
                            browser = FALSE) {
+  # Ensures that ARTnetData is installed
+  if (system.file(package = "ARTnetData") == "") stop(missing_data_msg)
 
   # Fix global binding check errors
   duration.time <- anal.acts.time <- anal.acts.time.cp <- NULL
@@ -124,8 +126,8 @@ build_epistats <- function(geog.lvl = NULL,
 
 
   ## Data ##
-  d <- ARTnet.wide
-  l <- ARTnet.long
+  d <- ARTnetData::ARTnet.wide
+  l <- ARTnetData::ARTnet.long
 
   out <- list()